Conclusão de todas as funções do sistema, aguardando testes...

+++ mvc/src/views/pages/filter.php
@@ -76,13 +76,14 @@
                             <td><?=$student->situation?></td>
                             <td>
                                 <div class="section-action-buttons">
-                                    <a href="<?=$base;?>/<?=$student->id?>/return" class="action-button">
+                                    <!--Função removida a pedido do usuário-->
+
+                                    <!--<a href="<?=$base;?>/<?=$student->id?>/return" class="action-button">
                                         <p>Retorno</p> <i class='bx bx-check act-icon'></i>
                                     </a>
+                                    /-->
 
-                                    /
-
-                                    <a href="<?=$base;?>/<?=$student->id?>/delete" class="action-button">
+                                    <a href="<?=$base;?>/<?=$student->id?>/delete" class="action-button" onclick='return confirmDel()'>
                                         <p>Excluir</p> <i class='bx bx-x act-icon'></i>
                                     </a>
                                 </div>
