+++ mvc/src/views/pages/home.php
@@ -2,7 +2,6 @@
 
     <section class="home">
         <h1>Saídas dos Alunos</h1>
-        
         <div class="section-area-add-student">
             <div class = "header-buttons">
                 <button class="add-button" onclick="openModal()">Adicionar</button>
@@ -73,13 +72,14 @@
                             <td><?=$student->situation?></td>
                             <td>
                                 <div class="section-action-buttons">
-                                    <a href="<?=$base;?>/<?=$student->id?>/return" class="action-button">
+                                    <!--Função removida a pedido do usuário-->
+
+                                    <!--<a href="<?=$base;?>/<?=$student->id?>/return" class="action-button">
                                         <p>Retorno</p> <i class='bx bx-check act-icon'></i>
                                     </a>
+                                    /-->
 
-                                    /
-
-                                    <a href="<?=$base;?>/<?=$student->id?>/delete" class="action-button">
+                                    <a href="<?=$base;?>/<?=$student->id?>/delete" class="action-button" onclick='return confirmDel()'>
                                         <p>Excluir</p> <i class='bx bx-x act-icon'></i>
                                     </a>
                                 </div>
